requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
jinja2==3.1.3
aiofiles==23.2.1
httpx==0.26.0
pillow==10.2.0
numpy==1.26.3
librosa==0.10.1
soundfile==0.12.1
transformers==4.37.2
torch==2.5.1+cpu
torchaudio==2.5.1+cpu
torchvision==0.20.1+cpu
opencv-python-headless==4.9.0.80
python-dotenv==1.0.0
scipy==1.12.0
matplotlib==3.8.2
PyWavelets==1.5.0
